double integration

+++ app/src/main/java/com/example/jonasexpensetracker/MainViewModel.java
@@ -7,27 +7,27 @@ import androidx.lifecycle.ViewModel;
 public class MainViewModel extends ViewModel {
 
   private final Model model;
-    private final MutableLiveData<Integer>monthlyIncomeLiveData=new MutableLiveData<>(0);
-    private final MutableLiveData<Integer>totalExpenseLiveData=new MutableLiveData<>(0);
-    private final MutableLiveData<Integer>currentBalanceLiveData=new MutableLiveData<>(0);
+    private final MutableLiveData<Double> monthlyIncomeLiveData = new MutableLiveData<>(0.0);
+    private final MutableLiveData<Double>totalExpenseLiveData=new MutableLiveData<>(0.0);
+    private final MutableLiveData<Double>currentBalanceLiveData=new MutableLiveData<>(0.0);
 
-    int inputNewExpense;
+    double inputNewExpense;
     public MainViewModel(Model model){
         super();
         this.model=model;
     }
 
-    public LiveData<Integer>getMonthlyIncome(){
+    public LiveData<Double>getMonthlyIncome(){
         return monthlyIncomeLiveData;
     }
-    public LiveData<Integer> getTotalExpense(){
+    public LiveData<Double> getTotalExpense(){
         return totalExpenseLiveData;
     }
-    public LiveData<Integer>getCurrentBalance(){
+    public LiveData<Double>getCurrentBalance(){
         return currentBalanceLiveData;
     }
 
-    public void getTotalExpenseFromModel(int amount){
+    public void getTotalExpenseFromModel(double amount){
         model.addExpenses(amount);
         totalExpenseLiveData.setValue(model.getTotal());
     }
@@ -38,10 +38,10 @@ public class MainViewModel extends ViewModel {
 
     public void onInputAmountChanged(String string){
         if(!string.equals("")) {
-            this.inputNewExpense = Integer.parseInt(string);
+            this.inputNewExpense = Double.parseDouble(string);
         }
     }
-    public int getInputAmount(){
+    public double getInputAmount(){
         return  inputNewExpense;
     }
